SIDM-9020-Added Accessibility tests.
