+++ src/test/a11y/aggregateAccessibilityResults.js
@@ -6,6 +6,28 @@ const outputFile = 'functional-output/accessibility/all_results.html';
 
 function aggregateAccessibilityResults() {
 
+
+
+    const regexPattern = /manage-user\.html$/;
+
+
+    findAndModifyHTML(regexPattern, 'New Heading Test');
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
     fs.readdir(htmlFilesDir, (err, files) => {
         if (err) {
             console.error('Error reading from accessibility results directory - ', err);
@@ -30,5 +52,51 @@ function aggregateAccessibilityResults() {
         });
     });
 }
+
+
+function findAndModifyHTML(regexPattern, newHeading) {
+    const directory = 'functional-output/accessibility';
+  
+    console.log('I am here 000000000.........');
+  
+    fs.readdir(directory, (err, files) => {
+      if (err) {
+        console.error('Error reading directory amend heading:', err);
+        return;
+      }
+  
+      const filenamePattern = new RegExp(regexPattern);
+  
+      files.forEach(filename => {
+        if (filenamePattern.test(filename)) {
+  
+          console.log('I am here 1111111.........');
+          const filePath = path.join(directory, filename);
+          fs.readFile(filePath, 'utf8', (err, data) => {
+            if (err) {
+              console.error(`Error reading file change heading ${filename}:`, err);
+              return;
+            }
+  
+            console.log('I am here 222222222.........');
+  
+            const modifiedContent = data.replace(/<h3>.*?<\/h3>/s, `<h3>${newHeading}</h3>`);
+  
+            console.log('I am here 3333333333.........');
+  
+            fs.writeFile(filePath, modifiedContent, 'utf8', err => {
+  
+              console.log('I am here 444444444.........');
+              if (err) {
+                console.error(`Error writing to file ${filename}:`, err);
+                return;
+              }
+              console.log(`Modified heading in ${filename}`);
+            });
+          });
+        }
+      });
+    });
+  }
 
 aggregateAccessibilityResults();
