+++ src/test/v2/accessibility/accessibility_test.ts
@@ -1,5 +1,5 @@
-const fs = require('fs');
-const path = require('path');
+// const fs = require('fs');
+// const path = require('path');
 
 Feature('v2_accessibility_tests');
 
@@ -21,10 +21,7 @@ Scenario('I am on manage user page',  async ({ I }) => {
 
   I.runA11yCheck({ reportFileName: 'manage-user.html' });
 
-  const regexPattern = /manage-user\.html$/;
 
-
-  findAndModifyHTML(regexPattern, 'New Heading Test');
 
   I.checkA11y();
 
@@ -63,47 +60,47 @@ Scenario('I am on manage user page',  async ({ I }) => {
 // });
 
 
-function findAndModifyHTML(regexPattern, newHeading) {
-  const directory = 'functional-output/accessibility';
+// function findAndModifyHTML(regexPattern, newHeading) {
+//   const directory = 'functional-output/accessibility';
 
-  console.log('I am here 000000000.........');
+//   console.log('I am here 000000000.........');
 
-  fs.readdir(directory, (err, files) => {
-    if (err) {
-      console.error('Error reading directory amend heading:', err);
-      return;
-    }
+//   fs.readdir(directory, (err, files) => {
+//     if (err) {
+//       console.error('Error reading directory amend heading:', err);
+//       return;
+//     }
 
-    const filenamePattern = new RegExp(regexPattern);
+//     const filenamePattern = new RegExp(regexPattern);
 
-    files.forEach(filename => {
-      if (filenamePattern.test(filename)) {
+//     files.forEach(filename => {
+//       if (filenamePattern.test(filename)) {
 
-        console.log('I am here 1111111.........');
-        const filePath = path.join(directory, filename);
-        fs.readFile(filePath, 'utf8', (err, data) => {
-          if (err) {
-            console.error(`Error reading file change heading ${filename}:`, err);
-            return;
-          }
+//         console.log('I am here 1111111.........');
+//         const filePath = path.join(directory, filename);
+//         fs.readFile(filePath, 'utf8', (err, data) => {
+//           if (err) {
+//             console.error(`Error reading file change heading ${filename}:`, err);
+//             return;
+//           }
 
-          console.log('I am here 222222222.........');
+//           console.log('I am here 222222222.........');
 
-          const modifiedContent = data.replace(/<h3>.*?<\/h3>/s, `<h3>${newHeading}</h3>`);
+//           const modifiedContent = data.replace(/<h3>.*?<\/h3>/s, `<h3>${newHeading}</h3>`);
 
-          console.log('I am here 3333333333.........');
+//           console.log('I am here 3333333333.........');
 
-          fs.writeFile(filePath, modifiedContent, 'utf8', err => {
+//           fs.writeFile(filePath, modifiedContent, 'utf8', err => {
 
-            console.log('I am here 444444444.........');
-            if (err) {
-              console.error(`Error writing to file ${filename}:`, err);
-              return;
-            }
-            console.log(`Modified heading in ${filename}`);
-          });
-        });
-      }
-    });
-  });
-}
+//             console.log('I am here 444444444.........');
+//             if (err) {
+//               console.error(`Error writing to file ${filename}:`, err);
+//               return;
+//             }
+//             console.log(`Modified heading in ${filename}`);
+//           });
+//         });
+//       }
+//     });
+//   });
+// }
